requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
distlib==0.3.4
filelock==3.7.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
googlemaps==4.6.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
platformdirs==2.5.2
psycopg2==2.9.3
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
strgen==1.3.1
urllib3==1.26.10
virtualenv==20.14.1
Werkzeug==2.1.2
zipp==3.8.0
